content/serializers.py
from rest_framework import serializers
from .models import Product, Lesson

class LessonSerializer(serializers.ModelSerializer):
    class Meta:
        model = Lesson
        fields = ('title', 'video_url')


class ProductSerializer(serializers.ModelSerializer):
    lessons = LessonSerializer(many=True, read_only=True)

    class Meta:
        model = Product
        fields = ('id', 'author', 'title', 'started_at', 'price', 'lessons')
